Checking in a few changes. Firstly, the == operator for Cfgs, it works now. Secondly, I am fixing the wierd bug where in 2D scenarios a line extends into infinity and beyond. Third, edge thickness for GL is a size_t now. And should also have a default value set....in our case it will be 1.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@294 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ Plum/MapObj/Edge.h
@@ -38,7 +38,7 @@ namespace plum {
     /////////////////////////////////////////////////////////////////
       
       bool BuildModels(){ return true; }
-      void SetThickness(float); 
+      void SetThickness(size_t); 
       void Draw(GLenum _mode);
       void DrawSelect();
 
@@ -93,7 +93,7 @@ namespace plum {
       double m_weight;
       int m_id;
 
-      float m_edgeThickness; 
+      size_t m_edgeThickness; 
 
       CCfg::Shape m_cfgShape;
